Move header and total back to main `ProjectBudget` component
