+++ src/scenes/Projects/Budget/ProjectBudgetRecords.tsx
@@ -1,6 +1,3 @@
-import { makeStyles, Typography } from '@material-ui/core';
-import { Skeleton } from '@material-ui/lab';
-import { sumBy } from 'lodash';
 import { Column, Components } from 'material-table';
 import React, { FC, useMemo } from 'react';
 import { useCurrencyFormatter } from '../../../components/Formatters/useCurrencyFormatter';
@@ -10,18 +7,6 @@ import {
   ProjectBudgetQuery,
   useUpdateProjectBudgetRecordMutation,
 } from './ProjectBudget.generated';
-
-const useStyles = makeStyles(({ spacing }) => ({
-  header: {
-    margin: spacing(3, 0),
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'flex-end',
-  },
-  totalLoading: {
-    width: '10%',
-  },
-}));
 
 const tableComponents: Components = {
   // No toolbar since it's just empty space, we don't use it for anything.
@@ -43,13 +28,10 @@ interface ProjectBudgetRecordsProps {
 
 export const ProjectBudgetRecords: FC<ProjectBudgetRecordsProps> = (props) => {
   const { loading, budget } = props;
-  const classes = useStyles();
   const formatCurrency = useCurrencyFormatter();
   const [updateBudgetRecord] = useUpdateProjectBudgetRecordMutation();
 
   const records: readonly BudgetRecord[] = budget?.value?.records ?? [];
-
-  const budgetTotal = sumBy(records, (record) => record.amount.value ?? 0);
 
   const rowData = records.map<BudgetRowData>((record) => ({
     id: record.id,
@@ -90,42 +72,27 @@ export const ProjectBudgetRecords: FC<ProjectBudgetRecordsProps> = (props) => {
   );
 
   return (
-    <>
-      <header className={classes.header}>
-        <Typography variant="h2">Budget</Typography>
-        <Typography
-          variant="h3"
-          className={loading ? classes.totalLoading : undefined}
-        >
-          {!loading ? (
-            `Total: ${formatCurrency(budgetTotal)}`
-          ) : (
-            <Skeleton width="100%" />
-          )}
-        </Typography>
-      </header>
-      <Table
-        data={rowData}
-        columns={columns}
-        isLoading={loading}
-        components={tableComponents}
-        cellEditable={
-          budget?.canEdit
-            ? {
-                onCellEditApproved: async (newAmount, _, data) => {
-                  if (newAmount === blankAmount || newAmount === '') return;
-                  const input = {
-                    budgetRecord: {
-                      id: data.id,
-                      amount: Number(newAmount),
-                    },
-                  };
-                  await updateBudgetRecord({ variables: { input } });
-                },
-              }
-            : undefined
-        }
-      />
-    </>
+    <Table
+      data={rowData}
+      columns={columns}
+      isLoading={loading}
+      components={tableComponents}
+      cellEditable={
+        budget?.canEdit
+          ? {
+              onCellEditApproved: async (newAmount, _, data) => {
+                if (newAmount === blankAmount || newAmount === '') return;
+                const input = {
+                  budgetRecord: {
+                    id: data.id,
+                    amount: Number(newAmount),
+                  },
+                };
+                await updateBudgetRecord({ variables: { input } });
+              },
+            }
+          : undefined
+      }
+    />
   );
 };
